Total rewrite.
Lots of users complained they couldn't run the MongoDB test, and when I
came back to run them, I found out I couldn't get them to run *either*.
There's a bunch of indirection layers for running stuff locally and
using MongoDB-specific network fault injection, which I'm sure is great
and can come back some day, but for the time being, none of it worked,
and it was too complicated for me to debug.

So, here's a fresh implementation of a test, based on Elle's list-append
workload and the new composite nemesis in Jepsen. We run sharded
clusters with replica sets for configsvr and shards, as recommended by
the docs. Use the all-new sync Java driver that the docs recommend.

Wooo!